Finalize merges
