+++ tests/main_test.py
@@ -1,6 +1,7 @@
 import pandas as pd
 from pathlib import Path
 from connectoviz.plot_circular_connectome import plot_circular_connectome
+import matplotlib.pyplot as plt
 
 # Path to the current script
 SCRIPT_DIR = Path(__file__).resolve().parent
@@ -21,12 +22,13 @@ atlas_path = ATLAS_DIR / atlas_fname
 
 matrix_path = MAT_DIR / matrix_fname
 
-atlas_fname = r"fan2016/MNI152/space-MNI152_atlas-fan2016_res-1mm_dseg.csv"
-matrix_fname = r"fan2016.csv"
-# Now construct full paths
-atlas_path = ATLAS_DIR / atlas_fname
 
-matrix_path = MAT_DIR / matrix_fname
+# atlas_fname = r"fan2016/MNI152/space-MNI152_atlas-fan2016_res-1mm_dseg.csv"
+# matrix_fname = r"fan2016.csv"
+# # Now construct full paths
+# atlas_path = ATLAS_DIR / atlas_fname
+
+# matrix_path = MAT_DIR / matrix_fname
 
 # example uasge witg Connectome class:
 atlas_pd = pd.read_csv(atlas_path)
@@ -41,7 +43,7 @@ layout_dict = {
     "display_node_name": False,
     "display_group_name": True,
 }
-plot_circular_connectome(
+fig, ax = plot_circular_connectome(
     con_mat=con_mat,
     atlas=atlas_pd,
     metadata_df=None,
@@ -53,8 +55,9 @@ plot_circular_connectome(
     label="Label",
     roi_names="ROIname",
     tracks=["Yeo_7network"],
+    show_graph=True,
 )
-
+plt.show()
 
 # test huang2022
 # atlas_fname = r"huang2022/MNI152/space-MNI152_atlas-huang2022_res-1mm_dseg.csv"
@@ -77,24 +80,24 @@ plot_circular_connectome(
 #      roi_names="RegionName1",
 #  )
 
-# test schaefer2018
-atlas_fname = r"schaefer2018tian2020/MNI152/space-MNI152_atlas-schaefer2018tian2020_res-1mm_den-100_div-7networks_dseg.csv"
-matrix_fname = r"schaefer2018tian2020_100_7.csv"
-# Now construct full paths
-atlas_path = ATLAS_DIR / atlas_fname
-matrix_path = MAT_DIR / matrix_fname
-atlas_pd = pd.read_csv(atlas_path)
-con_mat = pd.read_csv(matrix_path, header=None).values
-plot_circular_connectome(
-    con_mat=con_mat,
-    atlas=atlas_pd,
-    metadata_df=None,
-    hemispheric_par=True,
-    include_other=True,
-    group_by="component",
-    display_group_names=True,
-    display_node_names=False,
-    label="index",  # Assuming 'index' is the label column in the atlas
-    roi_names="name",
-    # tracks=["network"]
-)
+# # test schaefer2018
+# atlas_fname = r"schaefer2018tian2020/MNI152/space-MNI152_atlas-schaefer2018tian2020_res-1mm_den-100_div-7networks_dseg.csv"
+# matrix_fname = r"schaefer2018tian2020_100_7.csv"
+# # Now construct full paths
+# atlas_path = ATLAS_DIR / atlas_fname
+# matrix_path = MAT_DIR / matrix_fname
+# atlas_pd = pd.read_csv(atlas_path)
+# con_mat = pd.read_csv(matrix_path, header=None).values
+# plot_circular_connectome(
+#     con_mat=con_mat,
+#     atlas=atlas_pd,
+#     metadata_df=None,
+#     hemispheric_par=True,
+#     include_other=True,
+#     group_by="component",
+#     display_group_names=True,
+#     display_node_names=False,
+#     label="index",  # Assuming 'index' is the label column in the atlas
+#     roi_names="name",
+#     # tracks=["network"]
+# )
